+++ VFM/app/src/main/AndroidManifest.xml
@@ -29,6 +29,7 @@
         <activity android:name=".StartUp"/>
         <activity android:name=".LetsDrive"/>
         <activity android:name=".PreviousDrives"/>
+        <activity android:name=".DriveInfo"/>
 
         <meta-data
             android:name="preloaded_fonts"
