+++ VFM/app/src/main/java/com/cmsc436/vehiclefinemaps/LetsDrive.kt
@@ -13,8 +13,19 @@ import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
+import java.text.SimpleDateFormat
+import java.util.*
+import java.util.concurrent.TimeUnit
 import java.util.jar.Manifest
 import kotlin.random.Random
+
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.FirebaseDatabase
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.database.DataSnapshot
+
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.ValueEventListener
 
 
 
@@ -24,6 +35,15 @@ class LetsDrive : AppCompatActivity(){
     lateinit var startBtn: Button
     lateinit var endBtn: Button
     private var buttonCount = 0
+
+    //David's variables
+    var startTime: Long =0
+    lateinit var dateOfDrive:String
+    lateinit var timeOfDrive:String
+
+    private var mDatabaseReference: DatabaseReference? = null
+    private var mDatabase: FirebaseDatabase? = null
+    private var mAuth: FirebaseAuth? = null
 
 
 
@@ -36,6 +56,10 @@ class LetsDrive : AppCompatActivity(){
 
         findViewById<Button>(R.id.end_button).visibility = View.GONE
         findViewById<Button>(R.id.start_btn).visibility = View.VISIBLE
+
+        mDatabase = FirebaseDatabase.getInstance()
+        mDatabaseReference = mDatabase!!.reference!!.child("Previous")
+        mAuth = FirebaseAuth.getInstance()
 
         buttonCount = 0
 
@@ -58,12 +82,15 @@ class LetsDrive : AppCompatActivity(){
             buttonCount++;
 
             if (buttonCount % 2 == 0) {
+
                 findViewById<Button>(R.id.end_button).visibility = View.GONE
                 findViewById<Button>(R.id.start_btn).visibility = View.VISIBLE
             } else {
                 findViewById<Button>(R.id.end_button).visibility = View.VISIBLE
                 findViewById<Button>(R.id.start_btn).visibility = View.GONE
             }
+            //set the start time of the duration
+            startTime=System.currentTimeMillis()
 
         }
 
@@ -74,11 +101,51 @@ class LetsDrive : AppCompatActivity(){
                 findViewById<Button>(R.id.end_button).visibility = View.GONE
                 findViewById<Button>(R.id.start_btn).visibility = View.VISIBLE
             } else {
+
                 findViewById<Button>(R.id.end_button).visibility = View.VISIBLE
                 findViewById<Button>(R.id.start_btn).visibility = View.GONE
             }
 
             //SAVE CURRENT DRIVE INFORMATION
+            //once drive is done add date and time of drive to list
+            // and then add it to the database
+
+
+            //this formats the duration, date and time calculated to string form after drive ends
+            val endTime=System.currentTimeMillis()
+            val duration=endTime-startTime
+            val durationFormat=String.format("%d min, %d sec",
+                TimeUnit.MILLISECONDS.toMinutes(duration),
+                TimeUnit.MILLISECONDS.toSeconds(duration) -
+                        TimeUnit.MINUTES.toSeconds(TimeUnit.MILLISECONDS.toMinutes(duration))
+            );
+            val date= SimpleDateFormat("M/dd/yyyy")
+            dateOfDrive=date.format(Date())
+            val time= SimpleDateFormat("hh:mm:ss")
+            timeOfDrive=time.format(Date())
+
+
+            //Firebase database format for Previous Drives
+            // Previous
+            //  |   (the previous drives table contains all the users and each of their drives)
+            //  |
+            //  - UserId1
+            //       |      (contains all drives from this user
+            //       |
+            //        - DriveId1
+            //              |
+            //              |   (contains date,time,duration of this particalar drive)
+            // etc....
+
+            //this creates a new previos drive when user hits end drive and adds it to database
+            val userId=mAuth!!.currentUser!!.uid
+            val currentUser=mDatabaseReference!!.child(userId)
+            //should create a unique key value
+            val driveId=currentUser.push().key as String
+
+            currentUser.child(driveId).child("date").setValue(dateOfDrive)
+            currentUser.child(driveId).child("time").setValue(timeOfDrive)
+            currentUser.child(driveId).child("duration").setValue(durationFormat)
 
 
         }
