+++ VFM/app/src/main/java/com/cmsc436/vehiclefinemaps/PreviousDrives.kt
@@ -3,14 +3,10 @@ package com.cmsc436.vehiclefinemaps
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ArrayAdapter
-import android.widget.Button
 import android.widget.ListView
-import android.widget.AdapterView
 import androidx.appcompat.app.AppCompatActivity
-import java.text.SimpleDateFormat
-import java.util.*
 import kotlin.collections.ArrayList
-import java.util.concurrent.TimeUnit
+
 
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
@@ -22,14 +18,10 @@ import com.google.firebase.database.ValueEventListener
 
 class PreviousDrives : AppCompatActivity() {
 
-    lateinit var dateOfDrive:String
-    lateinit var timeOfDrive:String
-    lateinit var driveStartedBtn: Button
-    lateinit var driveEndedBtn: Button
+
     lateinit var listItems: ArrayList<String>
     lateinit var adapter:ArrayAdapter<String>
     lateinit var listView: ListView
-    var startTime: Long =0
     lateinit var storedTime:String
     lateinit var storedDate:String
     lateinit var storedDuration:String
@@ -44,66 +36,10 @@ class PreviousDrives : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.previous_drives)
-        initialize()
         listView=findViewById(R.id.list) as ListView
-
-        adapter= ArrayAdapter(this,android.R.layout.simple_list_item_1,listItems)
+        adapter= ArrayAdapter(getApplicationContext(),android.R.layout.simple_list_item_1,listItems)
         listView.setAdapter(adapter)
-
-        driveStartedBtn=findViewById<Button>(R.id.btn_driveStart) as Button
-        driveEndedBtn=findViewById<Button>(R.id.btn_driveEnd) as Button
-
-        driveStartedBtn.setOnClickListener {
-            startTime=System.currentTimeMillis()
-
-        }
-
-        driveEndedBtn.setOnClickListener {
-            //once drive is done add date and time of drive to list
-            // and then add it to the database
-
-
-            //this formats the duration, date and time calculated to string form
-            val endTime=System.currentTimeMillis()
-            val duration=endTime-startTime
-            val durationFormat=String.format("%d min, %d sec",
-                TimeUnit.MILLISECONDS.toMinutes(duration),
-                TimeUnit.MILLISECONDS.toSeconds(duration) -
-                        TimeUnit.MINUTES.toSeconds(TimeUnit.MILLISECONDS.toMinutes(duration))
-            );
-            val date=SimpleDateFormat("M/dd/yyyy")
-            dateOfDrive=date.format(Date())
-            val time=SimpleDateFormat("hh:mm:ss")
-            timeOfDrive=time.format(Date())
-
-
-            //Firebase database format for Previous Drives
-            // Previous
-            //  |   (the previous drives table contains all the users and each of their drives)
-            //  |
-            //  - UserId1
-            //       |      (contains all drives from this user
-            //       |
-            //        - DriveId1
-            //              |
-            //              |   (contains date,time,duration of this particalar drive)
-            // etc....
-
-            //this creates a new previos drive when user hits end drive and adds it to database
-            val userId=mAuth!!.currentUser!!.uid
-            val currentUser=mDatabaseReference!!.child(userId)
-            //should create a unique key value
-            val driveId=currentUser.push().key as String
-
-            currentUser.child(driveId).child("date").setValue(dateOfDrive)
-            currentUser.child(driveId).child("time").setValue(timeOfDrive)
-            currentUser.child(driveId).child("duration").setValue(durationFormat)
-
-
-
-        }
-
-
+        initialize()
 
     }
